Moves agent implementation to hazelcast-build-utils
When the FinalRemovalAgent is attached at runtime,
it is possible that some classes may have already been
loaded, resulting in random test failures in
compatibility tests. The hazelcast-build-utils
artifact declares the agent in its manifest
so it can be used with java -javaagent on the
command line.
